SequencePointComparer Test
