+++ main/OpenCover.Test/OpenCover.Test.csproj
@@ -167,6 +167,7 @@
     <Compile Include="Framework\Symbols\CecilSymbolManagerTests.cs" />
     <Compile Include="Framework\Utility\PerfCountersTests.cs" />
     <Compile Include="Framework\Utility\CodeCoverageStringTextSourceTest.cs" />
+    <Compile Include="Framework\Utility\SequencePointComparerTest.cs" />
     <Compile Include="Integration\SimpleBranchTests.cs" />
     <Compile Include="Integration\SimpleExceptionTests.cs" />
     <Compile Include="Integration\ThreadingTests.cs" />
